Added OnClick calls and fixed button layout.
Buttons from content_main now call methods in MainActivity when pressed.
These then call methods that will be implemented to actually send the
command to the UASC.
Before, the Emergency button was much wider than the other commands. A
minimum width has been added to the command buttons/switches. This helps
make all of the commands equal in width. This, however, causes issues on
screens with lower resolution. command_grid_layout has also
been centered horizontally to help center all of the burrons.
The light button has been changed to a switch to help the user tell if
the light is on or off.
Basic debug logging has also been added to help identify when buttons
are pressed, and what the new value of the light switch when it's
pressed.