Add fullscreen activity for UAS images
When the user presses the uas_image_view from the content_main layout,
a new activity will start. This activity goes fullscreen and is in
landscape orientation. Basic image conversions have been added (but are
NOT complete) that will later update the fullscreen_uas_image_view when
the app gets a new image from the UASC.