Add CommandService and communication intents
CommandService is responsible for sending new UAS images to the app
activities and handling commands.
When the user issues a command from the HHMD, that Activity sends an
Intent to the CommandService. Once receiving this Intent, the
CommandService will call the appropriate method to send the command to
the UASC. (command methods in the CommandService not yet implemented)
To act on information sent from the UASC (imagery and location) the
CommandService will receive an Intent from the thread that reads the
information (Not yet implemented). After reading and processing this,
the CommandService sends out a Broadcast. All Activities registered to
receive these Broadcasts should take action once receiving the
broadcast.
Since the Service should be running through the life of the application,
the CommandService is started in the SplashActivity. It may be ideal in
the future to have the SplashActivity wait until the CommandService is
created before transitioning to the MainActivity.
TempConstants has been removed since it is now unnecessary.
Three images have been added to the project. These are mainly for
debugging and testing. One temporary test can be done by toggling the
light switch multiple times. After toggling the switch twice, the
uas_image_view will cycle through the three different images.
FullscreenUasImageViewActivity now displays the current image.