Partially implements CloudRail
As of now only connection to DropBox is provided. To access this simply click "Cloud Storage"
on the menu bar and it will add a folder to your DropBox. If this doesn't work then I might of
only implemented it in a way that only adds it to my DropBox.