Fix spelling errors and CommandService intent handling bug
A string value for one of the buttons was accidentally spelled as
"botton". This commit changes the spelling of that in both the
content_main layout and the strings value file
The main switch statement in CommandService was missing a
break, this would cause the handleCommandHhmdEmergency method to be
called each time a new image was received.