Implemented the SettingsActivity and moved the cloud connection into it
The cloud handling was moved from the MainActivity to the SettingsActivity and the CommandServices.
New constants were added to the CommandService to allow for the settings application
to communicate with it. The methods for handling the communication between the SettingsActivity
and the CommandService have been added in however, note that only the method for changing
the cloud service being used is implemented.
NOTE:
The communication is set up but broadcasting from the SettingsActivity is not getting
to the CommandService for some reason.