Add session start preferences to new settings
Some of the settings that the user will be able to change have to do
with the start of a new flight session. This commit adds the header for
the session start preferences. Inside this header contains the UAS start
height preference. We may add additional settings for this header in the
future.