Add Wifi P2P activity to project
We are going to experiment some with P2P Wifi for connecting the HHMD to
the UASC. This is an activity that will display visible devices to the
user from which the user can select their UASC to connect to.
This commit simply adds the activity and some of it's resource files.