Prepare WifiP2pConnectActivity for WifiP2P content
This adds a BroadcastReceiver to receive intents from WifiP2pManager.
This will be used to update the list of nearby visible devices that are
available to connect to.
A ListView has been added to content_wifi_p2p_connect to show the nearby
devices to the user. The wifi_p2p_device layout will show basic
information about an individual visible nearby device.
SplashActivity has been modified to launch the WifiP2pConnectActivity
instead of the MainActivity after finishing the animation.