Add ProgressBar for loading animation
A spinning animation has been added to the Wifi Connect activity. This
will let the user know that the app is searching for nearby devices.
When the ListView gets an update/data change, a method gets called that
updates the visibility of the ProgressBar. That way, when results appear
the ProgressBar becomes invisible.