Add more WifiP2pManager content
This starts impementing the Wifi P2P content for the CommandService. On
creating the CommandService, a WifiP2pManager gets initializd and set up
and starts searching to discover peers.
For the WifiP2pConnectActivity, the ArrayAdapter variable is now global,
and the listview has a listener for its items. When Wifi P2p peers
change a new method gets called to add the new list of visible devices
to the ListView. This has not been fully tested and will need to be
worked on in the future.
The new required permissions for using Wifi P2p have been added to the
AndroidManifest.