Add WifiP2p device connection functionality
Clicking on a device listed in WifiP2pConnectActivity brings up a dialog
box asking if the user wishes to connect to that device. If the user
confirms the request, then an intent is sent to the CommandService. Upon
receiving this intent, the CommandService attempts to connect to the
device included in the intent.