Add location listener and location updates
Location is a key part of the app. It is essentially what makes the
whole app possible. This commit adds a location listener to provide the
HHMD's location to the CommandService every 3 seconds (this can be
changed by simply changing the new CONSTANT_LOCATION_UPDATE_SECONDS
constant) The CommandService registers to receive these location updates
after the session_start_button has been pressed. This can be changed
in the future if we want to receive location updates sooner.
The constants COMMAND_HHMD_LOCATION has been removed. This is because we
implement a listener that contains a method that gets called whenever
there is an update to the location. This method replaces the purpose
that the constant had in the broadcast receiver.