Add height offset to waypoint
When the user wants the UAS to fly higher or lower, they adjust the
height offset. When generating waypoints, it adds the height offset to
the altitude then clears the offset (since it has then been adjusted).
This ensures that the UAS is flying at the height that the user wants.
When a session starts, the height offset is set to the height set in the
UAS start height preference. If the user clicks on the height up or down
buttons on the MainActivity, the height offset gets incremented or
decremented respectively. Each time the height offset gets applied to a
newly generated waypoint, the offset is cleared out. So the height
offset is essentially the height vector that will get applied to the
next waypoint.

As of now we apply the diff vector and the height offset to the most
recent UAS location received. In the future we may want to apply those
changes to the most recent waypoint sent. This is because the UAS may
not have arrived at the previously sent waypoint when sending the HHMD
the UAS's location. This may lead to drift and inaccurate creation
of future waypoints. We may have to modify some of the logic of creating
waypoints in the future.

NOTE: This feature has not been tested because we have not yet connected
to and tested with the UAS.