Remove old settings activity, minor bug fixes
Since we are going to use ANdroid's Preferences to store app settings,
we no longer need the old SettingsActivity. Because of this, that
activity has been removed, and the NewSettingsActivity has been renamed
to SettingsActivity.

Some minor bug fixes have been added to keep the app from crashing on
startup, and throwing exceptions.