Implements saving of the Emergency Contacts list in the settings and using them
This Implements saving of the Emergency Contacts list by keeping a public static Preference Screen
over saving the data in something else like Preferred settings which might be a better way.
This also adds function to the Emergency button by logging all of the selected contacts and their
number to show that it is possible to be used for generating messages.