Add contacts selection list to emergency settings
This is a possible alternative to the existing contacts list. It scans
for contacts and adds them to the MultiSelectListPreference dialog box
for the user to select. As of now, this does not show the contact icon,
but this may be added later if we do choose to use this method of
listing contacts for the user.