Service binding initialization and Creation of UasCommunicationService
The purpose of the CommandService is to delegate tasks and actions
to other services, and be a central point for broadcasting intents to
activities. This will help keep the app code easier to read, debug, and
add features.

Activities that need to communicate with the CommandService will need to
bind to the CommandService (not required while still supporting
prototype intent handling) and make calls directly to it.
Activities will receive information from the CommandService through
broadcasted intents.

Services bound to the CommandService will make static calls to methods
within CommandService to broadcast intents to Activities

This commit creates one of the services aside from CommandService.
The UasCommunicationService is in charge of scanning and listening for
nearby WifiP2p devices, and connecting to a specified WifiP2p device.