Fix bug with Wifi P2P connections
For using the emulator, we need to get past the WifiP2PConnectActivity.
We did this by having it launch the MainActivity when attempting to
connect and receiving the P2P_UNSUPPORTED intent. Correctly launching
the MainActivity on a successful connection was accidentally left out.
This commit adds the MainActivity launch on a successful connection, and
distinguishes between the WifiP2p listeners for scanning and connecting.