Add LocationService for getting HHMD location and calculating waypoints
All location processing and listening that was in CommandService has
been pulled out into its own service.
When a new location for the HHMD has been processed, the LocationService
calculates a new waypoint to send to the UASC. After generating this, it
calls the CommandService statically which broadcasts an intent for the
CommandService containing the waypoint.

In the future, we might want to remove any large data from intents.
Instead of passing this data through intents, we should have the dynamic
methods get the data (that would otherwise be sent through the intents)
from the instances of the services. (e.g., The CommandService receives
an intent that a new Waypoint has been processed. The CommandService
then gets the latest waypoint from the LocationService and sends that
location to the UasCommunicationService)

Some minor log messages have been added and/or fixed