Change to disable/enable Start/End session buttons
This change was made for an easy way to prevent the application from
crashing if the End session button was clicked when a session was
currently not in progress.

There is still an issue where it is possible to spam the two buttons back
and forth which right now just creates a lot of folders in the users cloud.
However, this could be a big problem when it actually tells the UASC to
start/end a session.