Checking for null value and "None" cloud storage handling fixes
Two exceptions were commonly thrown because of null pointers.
When activities that show the UAS images are trying to get the latest
UAS image they call to get the local path to the latest image. If no
images have been received, a null value is returned. This null gets
treated as a path and throws a NullPointerException when this "path" is
opened. This is fixed by adding a simple check to make sure the path is
not null.

If no preference is stored for the cloud storage, getting the value for
the cloud storage preference is defaulted to a null value. This throws
an exception when this preference string is passed into a switch. This
commit sets the default value to pref_value_cloud_storage_provider_none
when no preference is set.