Small fix for permissions at the start of the application
The loop that looked for checking if a permission had been accepted or not
had a mistake in it where it was only looking at the first element. This would
cause problems where only 1 permission needed to be accepted and the others
would never be asked to be accepted if declined before.