Fix halting on SplashActivity
Issue #2 stated that in certain situations the SplashActivity would
not transition to WifiP2pConnectActivity. This would occur when the app
had successfully connected to a WifiP2p device, the user would close the
app, then disconnect from the WifiP2p device from Android's settings.
After this was done, whenever the app would be launched, it would not
transition past the SplashActivity.
This happened because the services would remain running after the app
was closed. When reopened, the services would already be bound to the
CommandService. Since they were already connected, it would broadcast
ACTION_UI_SERVICES_READY before the SplashActivity would get created.
Because the SplashActivity would never receive this broadcast, it would
not believe that the services were ready.
This was fixed for the SplashActivity by sending a broadcast asking
if the CommandService would broadcast if the services are ready.

This commit solves the issue on the UI side, but may need more
investigating for making sure that the services get stopped and
destroyed once the app has been closed.