Get Wifi P2p connection info on connection
In order to send messages, we need network information about connected
devices. This commit adds a Broadcast Receiver to get intents relating
to changes with Wifi P2p. When this receiver gets informed of a
connection change, it checks for the network information. If connected,
it will continue to the MainActivity.