Implements a basic Client connection to the selected Wi-Fi Direct Device
This implements basic connection to a selected Wi-Fi Direct Device acting as the server.
As of right now it only connects and sends the word "Ping".
It should also probably be moved out of where the UasClient is being created right now
such as not starting the connection until the MainActivity is created and disconnecting when
the application is closed.