WifiP2p disconnect and scan when HHMD initiates connection
Because of how the UASC works, we need to have the UASC establish the
connection with us. We do this by making an initial attempt (to give the
UASC our device information) then canceling the connection and waiting
for the UASC to connect with us. This also adds a UI dialog to let the
user know that the device is connecting instead of having no feedback.

This should resolve #4, but still needs to be thoroughly tested with a
device that has Wifi P2p to ensure it works properly.