Implements the beginning of the Heartbeat to the UASC
The heartbeat is to verify a constant connection to the UASC but before
this can be implemented completely a way to reestablish connection
when it is lost needs to be implemented.
Right now the only function it has is to send the server a "Ping" and then
check the response code. When the WifiNetworkState is not connected
it will just stop the heartbeat for now but will need to change the actual
function of how to handle disconnects later.