Implements the UASCClient for all connections to the server
This removes the Heartbeat class and adds it's functionality into the
new UASCClient class.

The UASCClient class has the ability to start the heartbeat and get the
latest image from the server. Accessing the latest gps needs to still
be implemented because the form it will be from the server is still unknown.