Rework CloudService to work with cloud storage or local saving
Because the user has the option to save images locally, we need a way to
handle this. This commit is used to handle the creation of directories,
images, and links for both cloud and local storage options. All of these
functions will be done on a separate thread so that it is not holding up
the main thread.

Some CommandService functionality has been changed so that it can get
newest data from the UasCommunicationService and pass it to the
CloudService.

NOTE: The method in UasCommunicationService has not been created or
implemented. This will be done once we determine what object type the
image data will be.