Implements accessing and updating the latest image from the server
This binds the MainActivity to the CommandService to be able to access the latest
image form the server.
Note: Many of the image saving methods need to be reworked and updated.

This implements the ability to access the latest image from the server but it for testing
the image being displayed is from an array list of URLs that contain an image.
The image updating is not started until the "take off" button is pressed.