Add new image uploading methods
Michael reworked some of the image uploading using the assumption that
the input will be a bitmap (commit f348d8e). This works better than
using a filename since it is not guaranteed that the image will be saved
as a file when a new image is received.
This commit takes that implementation and migrates it to the
cloud-service-rework branch.