Change UI image updating to use bound CommandService
Since broadcasts telling the UI that there is a new image will not
contain the image data or filename, the Activity must get the image data
a different way. This converts the old method of getting and setting the
image data to the ImageViews from using the data sent in the broadcasted
intent to using the bound CommandService object to retrieve the data.
Because of this change, the action String for broadcasting is no longer
used,. This commit cleans that up by removing it and using the static
calls to the CommandService for sending the broadcast to the UI.