Implement blank methods for accessing all server endpoints
This implements the blank methods for all of the planned server endpoints,
if more are needed they can easily be implemented.

It also fixes a few bugs that caused crashes:

handleNewImage() in onResume() in MainActivity and FullscreenUasImageActivity because
the image isn't available until the session is started so possibly add a blank image onto
the screens at that point.

UasCommunicationServiceBinder needs to be protected or the application won't compile
because another class needs access to it.