Add location calibration to LocationService
When the user starts a session, we want to wait until the locations are
"calibrated". Calibration is complete after we have received 3
sequential locations that have an accuracy less than 10 meters. When
this happens, a broadcast is sent to notify the UI that calibration is
complete. In the future, this notification to the UI will be used to
enable the user to begin issuing commands to the UAS.