Make services start not sticky
When the user closes the app after all of the services have started, it
appears that the UI is closed but the services continue to run. This is
problematic and causes some issues with state expectancy when the app is
restarted. This commit references #6 and maybe #2. We should test this
further to ensure that the services to not restart in different
situations.