Implements waiting on Location calibration and uploading images
Adds in the functionality so that the other start session methods will not be called
unless calibration is finished.
Note: Changed location minimum accuracy to 100 make it easier for testing

Adds in a way to stop the user from starting a new session until all of the images
are done being uploaded from the previous session. This is needed because
the phone cannot keep up with uploading images to the cloud every one second.