Add onLocationCalibrationComplete to starting services
Since we have a wait time for calibrating the location, we don't want
the services to perform all starting tasks when the startSession is
called. To fix this, these services have two methods that get called at
different states. The startSession methods get called when the user
presses the start button in the UI. This will be used to get things
ready for the new session. Once the location calibration is complete,
the services will have their onLocationCalibrationComplete method called
to notify them that the session is active.