MainActivity button state and calibrating dialog
While the LocationService is calibrating, we want to display some sort
of UI element so that the user doesn't think the app has froze. This
adds a dialog box that gets displayed when the user first presses the
"Take Off" button, and gets dismissed when calibration is complete.

Enabling and disabling buttons based off of app state has also been
added. This is currently being kept track in the MainActivity, so we may
want to move this to be in the CommandService in the future.