Image saving folder fixes and final path filename for upload/saving
There are multiple errors thrown about the file separator used in
defining the directory for saving images. This commit fixes that and
makes the session directory naming more consistent.

The path used for uploading/saving images locally has been pulled out of
the saveLocalImage and saveCloudImage methods and into the saveImage
method. This is because defining the path to save the image at was
redundant and could cause confusion. This also provides the opportunity
to make the path parameter final to ensure that it does not get changed
if a new session is started during an upload.