Add calls to show "Connecting..." dialog
All of the methods for displaying and dismissing the "Connecting..."
dialog were in place but were not called. This was because the initiator
variable was not set to HHMD when the connecting was instantiated by the
app. This commit sets the initiator, and uses it to broadcast when the
device is connecting or connected.