Clean up emergency message sending
This commit formats the message that will be sent out to the emergency
contacts. This happens in the case of the long message that sends the
location and cloud link in the message.
A Toast has also been added to be displayed on the MainActivity when the
emergency SMS has been sent.