Add local dispatch to top of emergency contact list
Since the main purpose of the emergency button is to notify local
dispatch/police, we need a way to always have their number in the list
of contacts. This enables the adding of the Local Emergency Dispatch to
the contacts.
NOTE: This is currently commented out so that we do not accidentally
text this number during testing.