Add state machine to CommandService
To keep with the model-view-controller paradigm, we need to be able to
see the state of the application from any connected Activity. This will
make it possible for the UI to react to different state changes within
the backend services. This commit adds the State object, and implements
the already existing UI responces to these state changes.