Add connecting dialog and dialog string resources
Currently, after a user selects to connect to a device, the confirmation
dialog is removed and nothing else is shown to let the user know that it
is connecting in the backgound. This commit adds a dialog box to be
shown to let the user know that the device is connecting.
This also makes use of string resources for dialog text, rather than
hard-coded string values.