Fix warnings with ArrayAdapter in WifiP2pConnectActivity
Listing the nearby Wifi devices to connect to uses an ArrayAdapter to
hold and displayed elements. Previously, the adapter declaration did not
state that the adapter would hold WifiP2pDevices, so some warnings were
shown to show that the Type had not been explicitly stated. This commit
fixes that warning so that the warning is no longer present.