Fix issue with creating local app folder
Starting a session while choosing to store locally caused an error with
creating the local app folder. This was caused because it was trying to
create a directory in the "/data" directory which is only writable by
the system. Since we weren't able to create the directory, any images
that we planned on storing would not be stored properly. This commit
changes the directory to store the app directory.