Add status text to dialog when connecting to UASC
Since we get notified when the connection state changes, we can display
that to the user so they know what is going on. This changes the single
"Connecting..." dialog to have three different messages. When the HHMD
makes the initial request to the UASC the dialog shows "Establishing
connection...". When the HHMD disconnects and is waiting for the UASC to
request to connect, the dialog shows "Waiting for reply...". Then, when
the UASC is initializing the connection with the HHMD, the dialog will
show "Connecting...". This sequence will give users a bit more context
and will help them understand why the connection process might be taking
longer than expected.