New way to access server images and bug fixes
This implements the new way of telling the server to rotate images before
accessing the image.

Fixes a bug where connecting to other devices that weren't the UASC wouldn't
continue to the MainActivity

Fixes a bug for duplicate names in the cloud because uploading as fast as once
a second sometimes causes this error.