Disable settings during a mission
During a session we do not want users to attempt to change some settings
(Cloud storage provider, Emergency contacts, etc.). One way to prevent
this is to only allow settings if the session state is SESSION_STOPPED.
If the session is not stopped and the user tries to go into settings, a
Toast message will appear to tell the user why they cannot access
settings.