Update ImageView as soon as connected to CommandService
In the MainActivity and the FullscreenUasImageActivity we want the
latest image from the server to be displayed. Currently, it waits until
a new image is received before updating the ImageView. We want this to
be more responsive when switching from the MainActivity to the
FullscreenUasImageActivity and not have as much of a delay. To do this,
we call to get the latest image as soon as the activity is bound to the
CommandService. This will reduce the delay making the transition between
activities much smoother.