Move settings headers into NavigationView
Instead of having a whole activity dedicated to holding three settings
headers, we could have a NavigationView hold those three headers. This
would make the settings more readily available from the MainActivity.
This adds four new Preference Activities, one of which is to perform
actions that are common to the other three to increase consistency.
This commit may help with issue #1.