Remove status bar throughout app
We want the user to be completely focused on the Help From Above app. To
do this better, we have all activities to be fullscreen to hide the
status bar. This also modifies the WifiP2pConnectActivity to use the
toolbar that is consistent with the rest of the app Activities.