Add getting contacts to EmergencyPreferenceActivity
Since moving to the drawer for holding the settings we have not had
getting the contacts for the user to select their Emergency Contacts.
This adds this into the EmergencyPreferenceActivity. This also adds
showing the emergency numbers selected in the preference summary. This
is a new feature but still needs some work as it only shows the phone
numbers and not the names (which would be preferred).