Dialog changes in MainActivity and WifiP2pConnectActivity
Previously, displaying the different messages on the dialog required
that the initial state in the sequence was passed through. This makes it
impossible to respond consistently if the state were to change in a
unpredicted manor. This is fine now, but we may want to be able to jump
into any state in the case that the user leaves the app then reenters.
This commit makes it possible for the dialog to be displayed regardless
of the state in transition. It also dismisses the dialog if the activity
is destroyed (if it is not dismissed, it seems to throw warnings/errors)