Fix "Emergency messages sent" Toast, enable/disable emergency button
The Toast for telling the user that the emergency messages have sent was
not being displayed in the MainActivity because the intent action was
not included in the IntentFilter. This fixes that, and fixes some
spelling in the Toast message. The MainActivity now handles emergency
session state for the emergency button, but it may need some sort of
delay before enabling (to prevent spamming the emergency button).