Implements sending the end session ,light toggle, and bug fix
Implements the endSessionRunnable to tell the server to end the
current session on its side.

Implements the light toggle to send to tell the server to turn on or
off the light but it is not used right now.

Fixes a rare bug where stopping the runnable that accesses the
images from the server wouldn't actually stop because if it is in
the middle of posting itself to the handler it would kill the main one
but still start up the one that was just posted.