Display image placeholder if not running a session
We had previously tried to display a basic image in the ImageViews if a
session is not running, but when we would call to refresh the image, it
would not select the right image because the Activity hadn't been bound
to the CommandService yet. This commit changes how we decide to display
the placeholder image, but using state logic when displaying a new
image. So if the app is running a session, it will get the latest UASC
image to display, otherwise, it will display the placeholder image.