UASCClient sendWaypoint and Thread separation
This implements sending the latest waypoint to the UASC.
Note: The waypoint being sent doesn't seem to be correct from testing.

To prevent any of the connections to the server from being blocked by
another message being sent, all of the handlers are on separate
threads now. The start and end session are on the same thread
because they should never be able to interfere with each other when
communicating to the server.