Adds notifying the command service with message status
This adds in telling the command service when all of the messages have
been sent as well as when all of the messages have been delivered.

Note 1:
No new session states where added because I don't understand those
at all. But to fix the emergency button from being active too quickly a state
for when all messages have been delivered can be added instead.

Note 2:
Disabled creating the cloud link  because it would cause crashes when
trying to test sending an emergency message.