Improve waypoint generation algorithm
It was noticed that the waypoints generated were inaccurate and needed
to be improved. I believe this is due to the difference in HHMD
locations is only taking into account the last two HHMD locations. This
can cause issues if more than two HHMD points are stored before we get
a UASC location. This commit should fix that issue by having temporary
Deques that store location data since the last waypoint. Once a waypoint
is generated, these deques are cleared so that we can see the changes
since the last waypoint generation.