Handle notifying CommandService if an image can't be saved locally
We had been ignoring the boolean result when creating a file to write
the image data to. This boolean is true when the file was successfully
created, and false if not. This commit uses this result to proceed or on
true, and throw an exception on false. When the methods catches an
exception, it logs the exception and statically calls the CommandService
to broadcast.
We now need to implement receiving this broadcast in the different
activities, and displaying a Toast or AlertDialog when this happens.