Clean up waypoint generation methods
To simplify some things, the pushing of new waypoints into the stack is
added into the generateWaypoint methods. Since we no longer need the
waypoint to be returned, we can simplify the logic in the method and
make it easier to read.
Some unused methods in LocationService have been removed.