Catch cloud storage authentication canceled and improve login
Previously, we had issues if the user selected to use a cloud storage
type but refused to log in. This scenario would send the app into a
state where it would crash immediately and whenever the user attempted
to relaunch the app. This commit adds a try-catch around the login to
see if the user was authenticated or if they canceled. If they
canceled, then the app will default to using local storage. This will
keep the app from being in the unstable state as described above.

Another login improvement added in this commit has to do with when the
user logs in to their specified could storage provider. Originally, this
request happened when the user launched the app. This caused issues
because the started activity would cover our app's permissions request
on startup. The improvements in this commit add the following new
session states:

        SESSION_NOT_PREPARED
        SESSION_PREPARING
        SESSION_READY

When the app is launched, it now starts in SESSION_NOT_PREPARED. When
the MainActivity is bound to the CommandService, it calls to start
preparing the services (which currently is only the CloudService) and
sets the state to SESSION_PREPARING. When the CloudService receives the
call to prepare itself, it then checks to see if the user has selected a
cloud storage provider in the preferences, and if so, it checks to see
if they are logged in. If not logged in then it displays the request
activity started by CloudRail. It then proceeds to set up the app
directory. If no cloud storage is used, then it creates the local
directory. Once this is all finished, the CloudService notifies the
CommandService that it is prepared and ready for a session.
This pushes the login request from the start of the app to before the
user starts a session. Which fixes some issues that are mentioned in #8.