Fix issue with CloudRail authentication loop
When MainActivity is resumed it prepares for a session which includes
making sure that the user has logged into the selected cloud storage.
This had caused an issue because after the user properly logged in, the
user would return to the MainActivity which restarted this preparing
sequence.
This commit adds state checking before calling to start preparing the
session.