Change local image saving to save to Pictures directory
When the user selected to use local image saving, the images had been
saved in the app data directory. This makes it so no other app can
access the images. This could be good for future versions where there is
a built in gallery to view different mission images. But for now, we
need for the user to be able to see the images in their standard
gallery.