Change language for local image saving
We had been using "None" as the preference used when the user wants to
save images on their device rather than on a cloud storage provider.
This commit changes the language of that preference to show "Local"
since "None" may lead the user to think that the images are not being
saved., which is not the case.