Handle emergency state if no emergency contacts are selected
If there are no emergency contacts selected in the users list of
emergency contacts, then we complete the emergency (since no messages
will be sent or delivered). This makes it so that the session is not
stuck in the emergency state waiting for messages to send since no
messages get sent if there are no emergency contacts.